uWebSockets.js ... installation is hell.
I spent all morning monkeying around with uWebSockets. Many frustrations just getting it installed. Tried compiling it manually, but node.js refused to recognize the module, said "module not found." Encountered the sociopolitical turmoil which drove the developer away from NPM.

Eventually I was forced to downgrade my node version to 13.14.0 in order to get the right ABI version

Seems to be functional now. I'll take a few whacks at getting a higher version, and hence a higher ABI, and hence a newer node version... but for now, here it is.